Script to fetch patches for stable release branches
Uses the `github` python3 module

    $ pip3 install --user pygithub
    $ ./api.py 25

Where '25' is the milestone number, f.ex.
https://github.com/mesonbuild/meson/milestone/25

Downloads each patch into ./patches, prefixed with a number that
should be followed when applying the patches.

To aid in the developer workflow, when you've applied patches, move
them to ./patches/done and they will be skipped when you run ./api.py
a second time.