Default to verifying that fixed issues on the milestone have PRs
This wasn't enabled by default because it is very slow, but
I generally forget to run this (I forgot for 0.55.1) so just make it
the default.

We can't add those PRs automatically because sometimes issues are
fixed without being automatically closed by a PR.

In general, the relationship between issues and PRs that fixed them is
very tenuous in the GitHub API. Which is also why this function is so
slow.