update patch generation to directly use the merged commit shas
Make use of cherry-pick + format-patch to generate the patches, and in
particular, use the -x flag to produce a paper trail of how these
commits were selected. This makes it easy to look at a backported commit
and see where it came from and what PR it was associated with.