port closed-via-PR detection to graphql
graphql feels extremely unpleasant to work with. But this info is
available there, and never has been for the REST api (why???) as
evidenced by the reliance on a crude hack, namely scanning events for a
commit reference "in the right vicinity".

And github stopped generating those reference events at all, so...